sketch out a single command
