+++ pkg/libratoaggregator/aggregator.go
@@ -7,8 +7,6 @@ import (
 	"log"
 	"net/http"
 	"time"
-
-	"github.com/canaryio/canary"
 )
 
 type gauge struct {
@@ -118,7 +116,7 @@ func (a *Aggregator) flush() error {
 	}
 
 	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("User-Agent", "canary/"+canary.Version)
+	req.Header.Add("User-Agent", "")
 
 	req.SetBasicAuth(a.User, a.Token)
