--- version.go
@@ -1,4 +1,0 @@
-package canary
-
-// the public version of canary
-const Version string = "v3"
